Delete chrome_extensions.js
This file is here to allow ClojureScript to use Google Closure advanced compilation option. But it is outdated and I'm going to archive this repo.

So if you want to use the code, you should get a fresh version in google/clojure-compiler externs folder, which you can find (along with some explanation), [in this StackOverflow question](https://stackoverflow.com/q/17757898/1814970).

Cheers!